refactor: 카테고리 이미지 컴포넌트 분리
