+++ src/components/restaurant/RestaurantItem.tsx
@@ -1,13 +1,8 @@
 import { Component } from 'react';
 import styled from 'styled-components';
-import asian from '../../asset/category-asian.png';
-import chinese from '../../asset/category-chinese.png';
-import korean from '../../asset/category-korean.png';
-import japanese from '../../asset/category-japanese.png';
-import western from '../../asset/category-western.png';
-import etc from '../../asset/category-etc.png';
 import { Restaurant } from '../../App';
 import { CATEGORIES } from './RestaurantList';
+import CategoryIcon from '../common/CategoryIcon';
 
 const SRestaurant = styled.li`
   display: flex;
@@ -15,54 +10,23 @@ const SRestaurant = styled.li`
 
   padding: 16px 8px;
 
-  border-bottom: 1px solid ${({ theme }) => theme.color.divideColor};
+  border-bottom: 1px solid var(--divide-color);
 
   cursor: pointer;
 `;
 
-const CategoryIcon = styled.div`
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  width: 64px;
-  height: 64px;
-  min-width: 64px;
-  min-height: 64px;
-
-  margin-right: 16px;
-
-  border-radius: 50%;
-  background: ${({ theme }) => theme.color.lightenColor};
-  img {
-    width: 36px;
-    height: 36px;
-  }
-`;
-
-const RestaurantInfo = styled.div`
+const Information = styled.div`
   display: flex;
   flex-direction: column;
   justify-content: flex-start;
   flex-grow: 1;
 `;
 
-const RestaurantName = styled.h3`
-  font-size: 18px;
-  line-height: 28px;
-  font-weight: 600;
+const Distance = styled.span`
+  color: var(--primary-color);
 `;
 
-const RestaurantDistance = styled.span`
-  font-size: 16px;
-  line-height: 24px;
-  font-weight: 400;
-  color: ${({ theme }) => theme.color.primaryColor};
-`;
-
-const RestaurantDescription = styled.p`
-  font-size: 16px;
-  line-height: 24px;
-  font-weight: 400;
+const Description = styled.p`
   display: -webkit-box;
 
   padding-top: 8px;
@@ -72,15 +36,6 @@ const RestaurantDescription = styled.p`
   -webkit-line-clamp: 2;
   -webkit-box-orient: vertical;
 `;
-
-const categoryIcon: Record<(typeof CATEGORIES)[keyof typeof CATEGORIES], string> = {
-  한식: korean,
-  중식: chinese,
-  일식: japanese,
-  양식: western,
-  아시안: asian,
-  기타: etc,
-};
 
 export type Props = {
   category: (typeof CATEGORIES)[keyof typeof CATEGORIES];
@@ -105,14 +60,12 @@ class RestaurantItem extends Component<Props> {
   render() {
     return (
       <SRestaurant onClick={this.onClickRestaurant}>
-        <CategoryIcon>
-          <img src={categoryIcon[this.props.category]} alt={this.props.category} />
-        </CategoryIcon>
-        <RestaurantInfo>
-          <RestaurantName>{this.props.name}</RestaurantName>
-          <RestaurantDistance>캠퍼스부터 {this.props.distanceByMinutes}분 내</RestaurantDistance>
-          <RestaurantDescription>{this.props.description}</RestaurantDescription>
-        </RestaurantInfo>
+        <CategoryIcon category={this.props.category} />
+        <Information>
+          <h3 className="text-subtitle">{this.props.name}</h3>
+          <Distance className="text-body">캠퍼스부터 {this.props.distanceByMinutes}분 내</Distance>
+          <Description className="text-body">{this.props.description}</Description>
+        </Information>
       </SRestaurant>
     );
   }
